fix: extract compute minimum mint into hyperwave
With the first refactor we turned the generic workflow
(that could run with or without computeMinimumMint), into
a modular workflow (that hardcoded each steps).

So computeMinimumMint did not need to be exposed into
the generic workflow, it just needs to be used by the
hyperwave-specific deposit function.

This simplifies the interfaces, and push back the Zapper to it's
new role: holding a bunch of data & accessors.